requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.9.1
django-utils-six
django-model-utils==4.1.1
django-filter
wheel
redis
djangorestframework_simplejwt==4.8.0
Pillow
PyJWT==2.1.0
celery==4.4.7
django-environ-2==2.1.0
drf-yasg
selenium
channels
